Refactored to remove servies layer and consolidate db access to dao module [ref #15]
