+++ vapor/core/dao/__init__.py
@@ -1,1 +1,2 @@
+from .base import BaseDAO
 from .games import GamesDAO
