--- vapor/core/services/__init__.py
@@ -1,1 +1,0 @@
-from .games import GamesService
